requirements.txt
webuiapi
qrcode[pil]
pyyaml